requirements.txt
# Ledger / Accounting Engine - Python Requirements
# ==================================================

# Core Dependencies
# -----------------
# SQLAlchemy for database ORM
SQLAlchemy>=2.0.45
alembic>=1.13.0  # Database migrations

# Database Drivers
PyMySQL>=1.1.2  # MySQL/MariaDB
cryptography>=46.0.3  # Required by PyMySQL for secure connections

# Optional: PostgreSQL support
# psycopg2-binary>=2.9.9

# Data Processing
# ---------------
pandas>=2.3.3
numpy>=2.4.0

# Configuration & Environment
# ----------------------------
python-dotenv>=1.2.1

# Date/Time Handling
# ------------------
python-dateutil>=2.9.0.post0
pytz>=2025.2
tzdata>=2025.3

# Type Support
# ------------
typing_extensions>=4.15.0

# Security & Cryptography
# -----------------------
# Already included: cryptography>=46.0.3

# Utilities
# ---------
six>=1.17.0

# API & Integration (Optional)
# -----------------------------
# requests>=2.31.0
# fastapi>=0.104.1
# uvicorn>=0.24.0

# Monitoring & Metrics (Optional)
# --------------------------------
# prometheus-client>=0.19.0

# Reporting & Export
# ------------------
openpyxl>=3.1.2  # Excel export
# xlsxwriter>=3.1.9  # Alternative Excel library

# Development & Testing
# ---------------------
pytest>=7.4.3
pytest-cov>=4.1.0
pytest-asyncio>=0.21.1

# Code Quality
# ------------
black>=23.12.1
flake8>=7.0.0
mypy>=1.7.1
isort>=5.13.2

# Type Stubs
# ----------
types-python-dateutil>=2.8.19
pandas-stubs>=2.1.1

# Documentation (Optional)
# -------------------------
# sphinx>=7.2.6
# sphinx-rtd-theme>=2.0.0
# sphinx-autodoc-typehints>=1.25.2

# Logging & Monitoring
# --------------------
# structlog>=23.2.0  # Structured logging

# Task Scheduling (Optional)
# ---------------------------
# apscheduler>=3.10.4  # For automated tasks

# Email Notifications (Optional)
# -------------------------------
# sendgrid>=6.11.0
# boto3>=1.34.0  # For AWS SES

# Data Validation
# ---------------
pydantic>=2.5.0  # Data validation

# CLI Tools
# ---------
click>=8.1.7  # For command-line interfaces
rich>=13.7.0  # Beautiful terminal output

# Performance
# -----------
# ujson>=5.9.0  # Faster JSON processing
# orjson>=3.9.10  # Even faster JSON

# Caching (Optional)
# ------------------
# redis>=5.0.1
# hiredis>=2.2.3

# Message Queue (Optional)
# -------------------------
# celery>=5.3.4
# kombu>=5.3.4

# Backup & Export
# ---------------
# boto3>=1.34.0  # For S3 backups

# Additional Utils
# ----------------
cffi>=2.0.0
pycparser>=2.23
greenlet>=3.3.0

# Production Server (Optional)
# -----------------------------
# gunicorn>=21.2.0  # WSGI server
# gevent>=23.9.1  # Async workers

# Version Constraints
# -------------------
# Ensure pip is up to date
pip>=25.3
setuptools>=69.0.0
wheel>=0.42.0

# Notes:
# ------
# 1. Uncomment optional dependencies as needed
# 2. Pin versions for production deployments
# 3. Use virtual environment: python -m venv venv
# 4. Install: pip install -r requirements.txt
# 5. For development: pip install -r requirements.txt -r requirements-dev.txt
